Update
Prefab updates (Updated tags, puzzle for all that needs activating, and button for the activators)
Wire Script so that you can draw in the wires as tilemaps to automatically connect up the different pieces.
A laser shooter to kill players and automatically getting the right lenght
Split that lets the player only controll one active clone at a time, and has a max amount of clones that is variable based on the level.
Start of the moving chamber script, which should allow the moving chamber to move to a set location